test/nba/test_expansion.py
import pandas
import pkg_resources
from unittest import TestCase

from dfs.nba.expansion import get_expansion_targets, encode_names, expand_nba_data, discretize_data

class ExpansionTestCase(TestCase):
  def setUp(self):
    # A little test data from the past few years, useful for testing BREF data
    testfn = pkg_resources.resource_filename(__name__, 'test.pickle')
    self.data = pandas.read_pickle(testfn)
    # More recent test data -- used for testing external data
    recentfn = pkg_resources.resource_filename(__name__, 'recent.pickle')
    self.recentdata = pandas.read_pickle(recentfn)
    # grab only one entry from each for super simple testing
    self.ezdata = pandas.concat([self.data.tail(1), self.recentdata.tail(1)])

  def test_get_expansion_targets(self):
    live_targets = list(get_expansion_targets(expanding_live=False))
    old_targets = list(get_expansion_targets())
    # Check types
    for name, (fn, targets) in live_targets + old_targets:
      assert isinstance(name, basestring)
      assert hasattr(fn, '__call__')
      assert isinstance(targets, list)
      for i in targets:
        assert isinstance(i, basestring)

  def test_encode_names(self):
    self.assertDictEqual({"A":"FeatureName:A", "B":"FeatureName:B", "C":"FeatureName:C"},
                         encode_names("FeatureName",["A","B","C"]))

  def test_expansion(self):
    # basically just make sure these don't crash
    expand_nba_data(self.data, live=False)
    expand_nba_data(self.recentdata, live=True)
    ez_expand = expand_nba_data(self.ezdata, live=False)
    ez_expand_live = expand_nba_data(self.ezdata, live=True)
    # this stuff sucks and was always getting killed b/c of data updates
    #self.maxDiff = None
    #count_dict = {'bref_id': 2, u'FT': 2, 'NF:STL': 1, 'OpponentLast2Weeks:MaxFPPMP': 2, u'3P': 2, u'TOV': 2, 'OpponentLast2Weeks:MaxFPPG': 2, u'Tm': 2, u'GmSc': 2, u'FG': 2, u'3PA': 2, u'DRB': 2, u'Rk': 2, 'NF:BLK': 1, u'Opp': 2, u'AST': 2, u'HomeAway': 0, u'FT%': 1, 'NF:Minutes': 1, u'PF': 2, 'NF:TOV': 1, u'PTS': 2, u'FGA': 2, 'Vegas:Spread': 2, 'OpponentLast2Weeks:AvgFPPG': 2, u'GS': 2, u'G': 2, 'NF:FP': 1, u'STL': 2, 'Last5:PPG': 2, u'Age': 2, u'TRB': 2, u'DFS': 1, u'FTA': 2, u'BLK': 2, 'date': 2, u'FG%': 2, 'OpponentLast2Weeks:AvgFPPMP': 2, 'Vegas:OverUnder': 2, u'+/-': 2, u'WinLoss': 2, 'NF:PTS': 1, 'Target:FDFP': 2, 'NF:REB': 1, 'NF:AST': 1, u'MP': 2, 'NF:PF': 1, 'OpponentLast2Weeks:FPPMP': 2, u'ORB': 2, u'3P%': 2, 'Salary:FanDuel Salary':1}
    #self.assertDictEqual(count_dict, ez_expand.count().to_dict())
    #live_count_dict = {'bref_id': 2, u'FT': 2, 'NF:STL': 1, 'OpponentLast2Weeks:MaxFPPMP': 2, u'3P': 2, u'TOV': 2, 'OpponentLast2Weeks:MaxFPPG': 2, u'Tm': 2, u'GmSc': 2, u'FG': 2, u'3PA': 2, u'DRB': 2, u'Rk': 2, 'NF:BLK': 1, u'Opp': 2, u'AST': 2, u'HomeAway': 0, u'FT%': 1, 'NF:Minutes': 1, u'PF': 2, 'NF:TOV': 1, u'PTS': 2, u'FGA': 2, 'Vegas:Spread': 2, 'OpponentLast2Weeks:AvgFPPG': 2, u'GS': 2, u'G': 2, 'NF:FP': 1, u'STL': 2, 'Last5:PPG': 2, u'Age': 2, u'TRB': 2, u'DFS': 1, u'FTA': 2, u'BLK': 2, 'date': 2, u'FG%': 2, 'OpponentLast2Weeks:AvgFPPMP': 2, 'Vegas:OverUnder': 1, u'+/-': 2, u'WinLoss': 2, 'NF:PTS': 1, 'NF:PF': 1, 'NF:REB': 1, 'NF:AST': 1, u'MP': 2, 'OpponentLast2Weeks:FPPMP': 2, u'ORB': 2, u'3P%': 2}
    #self.assertDictEqual(live_count_dict, ez_expand_live.count().to_dict())

  def test_discretization(self):
    stadium_series = pandas.Series(data=["Lambeau", "Levis", "Qwest"]) # Pretend this is an expanded field
    awesomeness_series = pandas.Series(data=[100, 30, 0]) # this is a continuous field
    name_series = pandas.Series(data=["Packers", "49ers", "Seahawks"]) # and this is a not-expanded discrete field
    df = pandas.DataFrame.from_dict({"Team:Stadium": stadium_series,
                                     "Awesomeness": awesomeness_series,
                                     "Team Name": name_series})
    discretized = discretize_data(df)
    # make sure only the expanded discrete fields were discretized
    self.assertItemsEqual(["Team:Stadium=Lambeau","Team:Stadium=Levis","Team:Stadium=Qwest","Awesomeness","Team Name"],
                          discretized.columns)